@@ -1,7 +1,7 @@
 package com.example.android.coordinatedeffort;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 
 public class SlidingCardsActivity extends AppCompatActivity {
