@@ -1,10 +1,9 @@
 package com.example.android.coordinatedeffort.behaviors;
 
 import android.content.Context;
-import android.support.design.widget.AppBarLayout;
-import android.support.design.widget.CoordinatorLayout;
+import com.google.android.material.appbar.AppBarLayout;
+import androidx.coordinatorlayout.widget.CoordinatorLayout;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.view.View;
 
 import com.example.android.coordinatedeffort.widget.FooterBarLayout;
